improve(patch): typings
