+++ config/tsconfig.json
@@ -21,15 +21,7 @@
     "skipLibCheck": true,
     "strict": false,
     "target": "es2021",
-    "types": [
-      "node",
-      "jest",
-      "react",
-      "express",
-      "webpack",
-      "@roots/bud",
-      "@roots/bud-framework"
-    ]
+    "types": ["node", "jest", "react", "express", "webpack"]
   },
   "files": [],
   "references": [
